requirements.txt
aiohttp==3.11.18
aiosmtplib==5.0.0
asyncpg==0.30.0
fastapi==0.124.4
httpx==0.28.1
matplotlib==3.10.8
openai==2.13.0
pandas==2.3.3
psycopg2_binary==2.9.11
pydantic==2.12.5
python-dotenv==1.2.1
python-telegram-bot==22.5
Requests==2.32.5
uvicorn==0.38.0
socksio
weasyprint
jinja2

pdfkit